feat(DatabaseFilmDirectorsStorage): optimize updateFilmDirectors() method by avoiding unnecessary operations

+++ src/main/java/ru/yandex/practicum/filmorate/storage/DatabaseFilmDirectorsStorage.java
@@ -39,8 +39,10 @@ public class DatabaseFilmDirectorsStorage {
 
     public void updateFilmDirectors(Film film) {
         if (film == null || film.getId() == null) throw new IllegalArgumentException("Фильм или ID фильма не могут быть null");
-        String sqlQuery = "DELETE FROM film_directors WHERE film_id = ?";
-        jdbcTemplate.update(sqlQuery, film.getId());
+        if (isFilmHasDirector(film.getId())) {
+            String sqlQuery = "DELETE FROM film_directors WHERE film_id = ?";
+            jdbcTemplate.update(sqlQuery, film.getId());
+        }
         if (film.getDirectors() != null && !film.getDirectors().isEmpty()) {
             saveFilmDirectors(film);
         }
